removed test package, moved decorator from nb test to nb test utils
